[Fixed] board

+++ webapp/src/main/java/com/doitmin/webapp/api/entities/Board.java
@@ -31,8 +31,8 @@ public class Board {
     private String deletedAt;
     private boolean isDeleted;
     private String imageUrl;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private String address;
     private String addressDetail;
     private String addressExtraInfo;
